+++ Sources/TunnelKitOpenVPNCore/ConfigurationParser.swift
@@ -111,6 +111,8 @@ extension OpenVPN {
             
             static let domain = NSRegularExpression("^dhcp-option +DOMAIN +[^ ]+")
             
+            static let domainSearch = NSRegularExpression("^dhcp-option +DOMAIN-SEARCH +[^ ]+")
+
             static let proxy = NSRegularExpression("^dhcp-option +PROXY_(HTTPS? +[^ ]+ +\\d+|AUTO_CONFIG_URL +[^ ]+)")
             
             static let proxyBypass = NSRegularExpression("^dhcp-option +PROXY_BYPASS +.+")
@@ -288,6 +290,7 @@ extension OpenVPN {
             var optRoutes4: [(String, String, String?)]?    // address, netmask, gateway
             var optRoutes6: [(String, UInt8, String?)]?     // destination, prefix, gateway
             var optDNSServers: [String]?
+            var optDomain: String?
             var optSearchDomains: [String]?
             var optHTTPProxy: Proxy?
             var optHTTPSProxy: Proxy?
@@ -657,6 +660,12 @@ extension OpenVPN {
                     guard $0.count == 2 else {
                         return
                     }
+                    optDomain = $0[1]
+                }
+                Regex.domainSearch.enumerateSpacedArguments(in: line) {
+                    guard $0.count == 2 else {
+                        return
+                    }
                     if optSearchDomains == nil {
                         optSearchDomains = []
                     }
@@ -931,6 +940,7 @@ extension OpenVPN {
             }
             
             sessionBuilder.dnsServers = optDNSServers
+            sessionBuilder.dnsDomain = optDomain
             sessionBuilder.searchDomains = optSearchDomains
             sessionBuilder.httpProxy = optHTTPProxy
             sessionBuilder.httpsProxy = optHTTPSProxy